Get Current Medicare Hospital Spending Data.

Get the current Medicare Hospital Spending data.

Usage

current_medicare_hospital_spending_data(
  .data,
  .data_sets = c("Claim", "Facility", "State", "National")
)

Arguments

.data

The data that results from the current_hosp_data() function.

.data_sets

The default is: c("Claim","Facility","State","National"), which will bring back all of the data in the Medicare Hospital Spending data sets that are in the file. You can choose from the following:

Claim

Facility

National

State

You can also pass things like c("state","Nation") as behind the scenes only the Medicare Hospital Spending data sets are available to the function to choose from and grep is used to find matches with ignore.case = TRUE set.

Value

Gets the current Medicare Hospital Spending data from the current hospital data file.

Details

This function will obtain the current Medicare Hospital Spending data from the output of the current_hosp_data() function, that is the required input for the .data parameter. You can pass in a list of which of those data sets you would like,

See also

Author

Steven P. Sanderson II, MPH

Examples

if (FALSE) {
current_hosp_data() |>
  current_maternal_data(.data_sets = c("State","National"))
}