Get Current PCH Oncology Measures Data.

Get the current PCH Oncology Measures data.

Usage

current_pch_oncology_measures_hospital_data(.data)

Arguments

.data

The data that results from the current_hosp_data() function.

Value

Gets the current PCH Oncology Measures Hospital data from the current hospital data file.

Details

This function will obtain the current PCH Oncology Measures data from the output of the current_hosp_data() function, that is the required input for the .data parameter. This function only returns a single object.

See also

Author

Steven P. Sanderson II, MPH

Examples

if (FALSE) { # \dontrun{
current_hosp_data() |>
  current_pch_oncology_measures_hospital_data()
} # }